+++ EssentialFeed/Image Comments API/ImageCommentsMapper.swift
@@ -7,7 +7,7 @@
 
 import Foundation
 
-enum ImageCommentsMapper {
+public enum ImageCommentsMapper {
     private struct Root: Decodable {
         private let items: [Item]
 
@@ -27,7 +27,7 @@ enum ImageCommentsMapper {
         }
     }
 
-    static func map(_ data: Data, from response: HTTPURLResponse) throws -> [ImageComment] {
+    public static func map(_ data: Data, from response: HTTPURLResponse) throws -> [ImageComment] {
         let decoder = JSONDecoder()
         decoder.dateDecodingStrategy = .iso8601
         guard isOK(response), let root = try? decoder.decode(Root.self, from: data)
